+++ CHANGELOG.rst
@@ -5,4 +5,5 @@ master
 ------
 
 - (`#35 <https://github.com/openclimatedata/openscm/pull/35>`_) Add units module
+- (`#44 <https://github.com/openclimatedata/openscm/pull/44>`_) Add timeframes module
 - (`#40 <https://github.com/openclimatedata/openscm/pull/40>`_) Add parameter handling in core module